change up widget verification components to fix text props
