+++ packages/mint-components/src/components/sqm-widget-verification/sqm-email-verification/readme.md
@@ -24,13 +24,13 @@
 ### Used by
 
  - [sqm-stencilbook](../../sqm-stencilbook)
- - [sqm-widget-verification-internal](..)
+ - [sqm-widget-verification](..)
 
 ### Graph
 ```mermaid
 graph TD;
   sqm-stencilbook --> sqm-email-verification
-  sqm-widget-verification-internal --> sqm-email-verification
+  sqm-widget-verification --> sqm-email-verification
   style sqm-email-verification fill:#f9f,stroke:#333,stroke-width:4px
 ```
