+++ packages/mint-components/src/components/tax-and-cash/sqm-banking-info-form/useBankingInfoForm.tsx
@@ -58,7 +58,6 @@ export type BankingInfoFormData = {
   paymentThreshold?: string;
   paymentSchedulingType?: "FIXED_DAY" | "BALANCE_THRESHOLD";
   paymentDay?: string;
-  email?: string;
 
   // Fields that aren't...
   beneficiaryAccountName?: string;
@@ -533,7 +532,7 @@ export function useBankingInfoForm(
       hasPayPal,
       bankCountry: formState.bankCountry,
       countrySearch,
-      email: userData.user.email,
+      email: userData?.user?.email,
     },
     refs: {
       formRef,
